Fixed a stupid bug causing the GraceClose() to kill all user processes if by that time the Grace subprocess has exited.

+++ grace_np/grace_np.c
@@ -320,7 +320,7 @@ GraceClose(void)
     }
 
     /* Tell grace to exit */
-    if (GraceCommand ("exit") == -1 || GraceFlush() == -1){
+    if (pid > 0 && (GraceCommand ("exit") == -1 || GraceFlush() == -1)) {
         kill(pid, SIGTERM);
     }
